requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
exceptiongroup==1.3.1
Faker==40.8.0
fastapi==0.135.1
h11==0.16.0
idna==3.11
numpy==2.2.6
pandas==2.3.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2026.1.post1
redis==7.3.0
requests==2.32.5
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
